Handle screens without mesh renderers during bake
This commit adds a small level of handling to an object with an
LTCGI_Screen and no mesh renderer during a bake. The bake proceeds but
the screen will not contribute any emissons. The user is warned and
pointed to the object that lacks the renderer via the debug logs.